move patch_ray_trafo_for_pickling from torch_utility to odl_utility
